# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da API cd_express",
    "version": "1.0.0",
    "description": "Documentação da API para o sistema CD Express"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "paths": {
...
    "/usuarios": {
      "get": {
        "summary": "Obter todos os usuários",
        "responses": {
          "200": {
            "description": "Sucesso ao obter todos os usuários"
          },
          "404": {
            "description": "Usuários não encontrados"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Criar um novo usuário",
        "responses": {
          "201": {
            "description": "Sucesso ao criar um novo usuário"
          },
          "400": {
            "description": "Solicitação inválida"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  }
}